feat: style 컴포넌트 관련 이름 변경 및 폴더 위치 변경, GlobalStyle 컴포넌트 생성
