--- frontend/tiggle/src/assets/base.css
@@ -1,5 +1,0 @@
-body {
-    margin: 0;
-    width: 100vw;
-    height: 100vh;
-}
