--- frontend/tiggle/src/components/atoms/headerRight.styled.ts
@@ -1,10 +1,0 @@
-import styled from "styled-components";
-
-export const StyledHeaderRight = styled.div`
-  display: flex;
-  align-items: center;
-
-  .bell {
-    margin-right: 1rem;
-  }
-`;
