+++ frontend/tiggle/src/components/organisms/Footer.tsx
@@ -1,13 +1,13 @@
 import { Award, Home, PieChart } from "react-feather";
-import { StyledFooter } from "./footer.styled";
+import { FooterStyle } from "../../styles/FooterStyle";
 import FooterIcon from "../atoms/FooterIcon";
 
 export default function Footer() {
   return (
-    <StyledFooter>
+    <FooterStyle>
       <FooterIcon icon={<Home />} iconName="홈" iconClass="focus" />
       <FooterIcon icon={<PieChart />} iconName="통계" />
       <FooterIcon icon={<Award />} iconName="랭킹" />
-    </StyledFooter>
+    </FooterStyle>
   );
 }
