+++ frontend/tiggle/src/components/organisms/Header.tsx
@@ -1,14 +1,13 @@
-import { StyledHeader } from "./header.styled";
+import { HeaderStyle } from "../../styles/HeaderStyle";
 import HeaderSection from "../molecules/HeaderSection";
-import "../../assets/base.css";
 import Navigation from "../molecules/Navigation";
 
 export default function Header() {
   return (
     <>
-      <StyledHeader>
+      <HeaderStyle>
         <HeaderSection />
-      </StyledHeader>
+      </HeaderStyle>
       {/* 768px 이하일 때부터 보이는 부분 */}
       <Navigation />
     </>
