+++ frontend/tiggle/src/index.tsx
@@ -5,7 +5,6 @@ import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 import { Provider } from "react-redux";
 import { BrowserRouter } from "react-router-dom";
 import { store } from "./store";
-import "./assets/base.css";
 
 const queryClient = new QueryClient();
 const container = document.getElementById("root");
