+++ frontend/tiggle/src/pages/Main.tsx
@@ -1,10 +1,12 @@
 import Header from "../components/organisms/Header";
 import Footer from "../components/organisms/Footer";
 import Content from "../components/organisms/Content";
+import { GlobalStyle } from "../styles/GlobalStyle";
 
 const Main = () => {
   return (
     <>
+      <GlobalStyle />
       <Header />
       <Content />
       <Footer />
