+++ frontend/tiggle/src/styles/GlobalStyle.ts
@@ -1,0 +1,9 @@
+import { createGlobalStyle } from "styled-components";
+
+export const GlobalStyle = createGlobalStyle`
+  body {
+    margin: 0;
+    width: 100vw;
+    height: 100vh;
+  }
+`;
