+++ frontend/tiggle/src/utils/global.d.ts
